# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/join-requests": {
      "get": {
        "operationId": "JoinRequestsController_getAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Role used by the RBAC guard",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "description": "Filter by project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filter by user ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of join requests returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JoinRequestEntity"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          }
        },
        "summary": "List all join requests",
        "tags": [
          "JoinRequests"
        ]
      },
      "post": {
        "operationId": "JoinRequestsController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Role used by the RBAC guard",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "user"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJoinRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Join request created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinRequestEntity"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          }
        },
        "summary": "Create a new join request",
        "tags": [
          "JoinRequests"
        ]
      }
    },
    "/join-requests/{id}": {
      "get": {
        "operationId": "JoinRequestsController_getOne",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Role used by the RBAC guard",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Join request ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Join request returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinRequestEntity"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Join request not found."
          }
        },
        "summary": "Get a join request by ID",
        "tags": [
          "JoinRequests"
        ]
      },
      "put": {
        "operationId": "JoinRequestsController_update",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Role used by the RBAC guard",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Join request ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJoinRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Join request updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinRequestEntity"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Join request not found."
          }
        },
        "summary": "Update an existing join request",
        "tags": [
          "JoinRequests"
        ]
      },
      "delete": {
        "operationId": "JoinRequestsController_remove",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Role used by the RBAC guard",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Join request ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Join request deleted successfully."
          },
          "403": {
            "description": "Forbidden - Admin role required"
          },
          "404": {
            "description": "Join request not found."
          }
        },
        "summary": "Delete a join request (Admin only)",
        "tags": [
          "JoinRequests"
        ]
      }
    },
...
  },
  "info": {
    "title": "TeamForge API",
    "description": "The Student Project Collaboration Platform REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "JoinRequestEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the join request"
          },
          "userId": {
            "type": "string",
            "description": "User ID making the request"
          },
          "userName": {
            "type": "string",
            "description": "User display name"
          },
          "projectId": {
            "type": "string",
            "description": "Target project ID"
          },
          "status": {
            "type": "string",
            "description": "Request status",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "message": {
            "type": "string",
            "description": "User's pitch/application message"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "userId",
          "userName",
          "projectId",
          "status",
          "message",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateJoinRequestDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "User ID making the request"
          },
          "userName": {
            "type": "string",
            "description": "User display name"
          },
          "projectId": {
            "type": "string",
            "description": "Target project ID"
          },
          "message": {
            "type": "string",
            "description": "User's pitch/application message"
          }
        },
        "required": [
          "userId",
          "userName",
          "projectId",
          "message"
        ]
      },
      "UpdateJoinRequestDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "User ID making the request"
          },
          "userName": {
            "type": "string",
            "description": "User display name"
          },
          "projectId": {
            "type": "string",
            "description": "Target project ID"
          },
          "message": {
            "type": "string",
            "description": "User's pitch/application message"
          },
          "status": {
            "type": "string",
            "description": "Request status",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          }
        }
      },
...
    }
  }
}